moved to sqlite

+++ apps/backend/modules/policies/policy.controller.js
@@ -1,3 +1,4 @@
+const constants = require("../../constants");
 const { extractError } = require("../../utils/error.utils");
 const Logger = require("../../utils/logger");
 const { PolicyService } = require("./policy.service");
@@ -14,10 +15,24 @@ policyController.getAllPolicies = async (req, res) => {
   try {
     const { pmUser, state, params } = req;
     const pm_user_id = parseInt(pmUser.pm_user_id);
+    const authorized_policies = state.authorized_policies;
     Logger.log("info", {
       message: "policyController:getAllPolicies:params",
       params: { pm_user_id },
     });
+    if (
+      !authorized_policies ||
+      (Array.isArray(authorized_policies) && authorized_policies.length == 0)
+    ) {
+      Logger.log("error", {
+        message: "policyController:getAllPolicies:catch-1",
+        params: { error: constants.ERROR_CODES.PERMISSION_DENIED },
+      });
+      return res.json({
+        success: false,
+        error: extractError(constants.ERROR_CODES.PERMISSION_DENIED),
+      });
+    }
     const policies = await PolicyService.getAllPolicies();
     Logger.log("success", {
       message: "policyController:getAllPolicies:success",
@@ -48,12 +63,24 @@ policyController.getPolicyByID = async (req, res) => {
     const { pmUser, state, params } = req;
     const pm_policy_object_id = parseInt(params.id);
     const pm_user_id = parseInt(pmUser.pm_user_id);
-
+    const authorized_policies = state.authorized_policies;
     Logger.log("info", {
       message: "policyController:getPolicyByID:params",
       params: { pm_user_id, pm_policy_object_id },
     });
-
+    if (
+      !authorized_policies ||
+      (Array.isArray(authorized_policies) && authorized_policies.length == 0)
+    ) {
+      Logger.log("error", {
+        message: "policyController:getPolicyByID:catch-1",
+        params: { error: constants.ERROR_CODES.PERMISSION_DENIED },
+      });
+      return res.json({
+        success: false,
+        error: extractError(constants.ERROR_CODES.PERMISSION_DENIED),
+      });
+    }
     const policy = await PolicyService.getPolicyByID({
       pmPolicyObjectID: pm_policy_object_id,
     });
@@ -86,7 +113,7 @@ policyController.addPolicy = async (req, res) => {
   try {
     const { pmUser, state, body } = req;
     const pm_user_id = parseInt(pmUser.pm_user_id);
-
+    const authorized_policies = state.authorized_policies;
     Logger.log("info", {
       message: "policyController:addPolicy:params",
       params: { pm_user_id, body },
@@ -125,11 +152,24 @@ policyController.updatePolicy = async (req, res) => {
   try {
     const { pmUser, state, body } = req;
     const pm_user_id = parseInt(pmUser.pm_user_id);
+    const authorized_policies = state.authorized_policies;
     Logger.log("info", {
       message: "policyController:updatePolicy:params",
       params: { pm_user_id, body },
     });
-
+    if (
+      !authorized_policies ||
+      (Array.isArray(authorized_policies) && authorized_policies.length == 0)
+    ) {
+      Logger.log("error", {
+        message: "policyController:updatePolicy:catch-1",
+        params: { error: constants.ERROR_CODES.PERMISSION_DENIED },
+      });
+      return res.json({
+        success: false,
+        error: extractError(constants.ERROR_CODES.PERMISSION_DENIED),
+      });
+    }
     const updatedPolicy = await PolicyService.updatePolicy({
       pmPolicyObjectID: parseInt(body.pm_policy_object_id),
       pmPolicyObjectTitle: body.pm_policy_object_title,
@@ -167,12 +207,24 @@ policyController.deletePolicy = async (req, res) => {
     const { pmUser, state, params } = req;
     const pm_policy_object_id = parseInt(params.id);
     const pm_user_id = parseInt(pmUser.pm_user_id);
-
+    const authorized_policies = state.authorized_policies;
     Logger.log("info", {
       message: "policyController:deletePolicy:params",
       params: { pm_user_id, pm_policy_object_id },
     });
-
+    if (
+      !authorized_policies ||
+      (Array.isArray(authorized_policies) && authorized_policies.length == 0)
+    ) {
+      Logger.log("error", {
+        message: "policyController:deletePolicy:catch-1",
+        params: { error: constants.ERROR_CODES.PERMISSION_DENIED },
+      });
+      return res.json({
+        success: false,
+        error: extractError(constants.ERROR_CODES.PERMISSION_DENIED),
+      });
+    }
     await PolicyService.deletePolicy({
       pmPolicyObjectID: pm_policy_object_id,
     });
